requirements.txt
streamlit==1.12.2
PyQRCode==1.2.1
pypng==0.20220715.0